app/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
sqlalchemy==2.0.30
python-multipart==0.0.9
